* A small hack for the compilation warning
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@3356 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ itemattributes.cpp
@@ -340,8 +340,8 @@ bool ItemAttribute::unserialize(PropStream& stream)
 		}
 		case FLOAT:
 		{
-			uint32_t v;
-			if(!stream.getLong(v))
+			float v;
+			if(!stream.getFloat(v))
 				return false;
 
 			set(*reinterpret_cast<float*>(&v));
